+++ .gitignore
@@ -11,3 +11,5 @@ instance/*
 __pycache__/
 *.py[cod]
 *$py.class
+
+.vscode
