--- README.md
@@ -1,5 +1,0 @@
-# Flask Middleware [WIP]
-
-this middleware/extension enables integrating UniAuth with flask applications effortless.  
-
-Read [how to develop this middleware here](https://github.com/UniAuth/UniAuth/wiki)
