+++ __init__.py
@@ -1,6 +1,6 @@
 from flask import Flask, Request, Response, redirect
 from config import Config
 from authenticate import AuthenticateParams
-from minions import getProfileData
+from app import getProfileData
 from urllib import request,parse
 import json
