--- src/interface/authenticate.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class authenticateParamsInterface:
-    def __init__(self):
-        self.successRedirect = str
-        self.failureRedirect = str
-AuthenticateParams = authenticateParamsInterface()
-
